create a global default icon instance and move related functions into a file with it
